+++ pkg/inject/injector.go
@@ -12,6 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// Package inject provides injection utilities.
+//
+// Deprecated: This package is no longer supported.
 package inject
 
 import (
@@ -24,8 +27,8 @@ import (
 
 	"github.com/cilium/ebpf"
 
-	"go.opentelemetry.io/auto/pkg/log"
-	"go.opentelemetry.io/auto/pkg/process"
+	"go.opentelemetry.io/auto/pkg/log"     // nolint:staticcheck  // Atomic deprecation.
+	"go.opentelemetry.io/auto/pkg/process" // nolint:staticcheck  // Atomic deprecation.
 )
 
 var (